% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeRNASeq.R
\name{normalizeRNASeq}
\alias{normalizeRNASeq}
\title{Normalize RNA-Seq data}
\usage{
normalizeRNASeq(geneFPKMFile = NULL, outputFolder = "./", name = NULL,
  groupString = NULL, useERCC = FALSE,
  erccFile = system.file("extdata/ERCC_Controls_Analysis.txt", package =
  "RNASeqT"))
}
\arguments{
\item{geneFPKMFile}{Cuffnorm results file (genes.fpkm_table)}

\item{outputFolder}{Output folder (default = './')}

\item{name}{Results file name prefix}

\item{groupString}{Comma-separated string of sample group names}

\item{useERCC}{Use ERCC spike-in normalization (default = FALSE)}

\item{erccFile}{erccFile}
}
\value{
None
}
\description{
Normalize RNA-Seq data
}
\author{
Charles Lin, \email{Charles.Y.Lin@bcm.edu}
}

